+++ pack-check.c
@@ -175,7 +175,7 @@ int verify_pack_index(struct packed_git *p)
 
 	/* Verify SHA1 sum of the index file */
 	git_SHA1_Init(&ctx);
-	git_SHA1_Update(&ctx, index_base, (unsigned int)(index_size - 20));
+	git_SHA1_Update(&ctx, index_base, (size_t)(index_size - 20));
 	git_SHA1_Final(sha1, &ctx);
 	if (hashcmp(sha1, index_base + index_size - 20))
 		err = error("Packfile index for %s SHA1 mismatch",
